modified:   src/App.js 	renamed:    src/pages/Navigation.js -> src/components/pages/Navigation.js 	renamed:    src/pages/StatedGraphForm.js -> src/components/pages/StatedGraphForm.js renamed:    src/pages/StatedNameForm.js -> src/components/pages/StatedNameForm.js
