+++ src/components/pages/StatedGraphForm.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Chart from "../components/Chart";
+import Chart from "../Chart";
 let GraphForm = () => {
   return (
     <div>
